Move routing interfaces to higher ktor abstraction level
