+++ tnoodle-server/src/main/kotlin/org/worldcubeassociation/tnoodle/server/routing/JsEnvHandler.kt
@@ -4,7 +4,7 @@ import io.ktor.application.call
 import io.ktor.http.ContentType
 import io.ktor.response.respond
 import io.ktor.response.respondText
-import io.ktor.routing.Routing
+import io.ktor.routing.Route
 import io.ktor.routing.get
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.builtins.serializer
@@ -13,7 +13,7 @@ import org.worldcubeassociation.tnoodle.server.RouteHandler
 import org.worldcubeassociation.tnoodle.server.serial.JsonConfig
 
 object JsEnvHandler : RouteHandler {
-    override fun install(router: Routing) {
+    override fun install(router: Route) {
         router.get("/env.js") {
             val serial = MapSerializer(String.serializer(), String.serializer())
             val js = "window.TNOODLE_ENV = ${JsonConfig.SERIALIZER.stringify(serial, JS_ENV)};"
