+++ tnoodle-server/src/main/kotlin/org/worldcubeassociation/tnoodle/server/routing/StylesheetHandler.kt
@@ -2,11 +2,11 @@ package org.worldcubeassociation.tnoodle.server.routing
 
 import io.ktor.http.content.resources
 import io.ktor.http.content.static
-import io.ktor.routing.Routing
+import io.ktor.routing.Route
 import org.worldcubeassociation.tnoodle.server.RouteHandler
 
 object StylesheetHandler : RouteHandler {
-    override fun install(router: Routing) {
+    override fun install(router: Route) {
         router.static("/css") {
             resources("css")
         }
