+++ tnoodle-server/src/main/kotlin/org/worldcubeassociation/tnoodle/server/routing/VersionHandler.kt
@@ -2,7 +2,7 @@ package org.worldcubeassociation.tnoodle.server.routing
 
 import io.ktor.application.call
 import io.ktor.response.respond
-import io.ktor.routing.Routing
+import io.ktor.routing.Route
 import io.ktor.routing.get
 import kotlinx.serialization.json.json
 
@@ -11,7 +11,7 @@ import org.worldcubeassociation.tnoodle.server.cryptography.AsymmetricCipher
 import org.worldcubeassociation.tnoodle.server.util.ServerEnvironmentConfig
 
 class VersionHandler(val version: ServerEnvironmentConfig) : RouteHandler {
-    override fun install(router: Routing) {
+    override fun install(router: Route) {
         router.get("version") {
             //FIXME this is a temporary stub implementation until we have actual key pairs
             //val buildVerified = BuildVerification.BUILD_VERIFIED
