+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/StaticContentHandler.kt
@@ -4,11 +4,11 @@ import io.ktor.http.content.defaultResource
 import io.ktor.http.content.resource
 import io.ktor.http.content.resources
 import io.ktor.http.content.static
-import io.ktor.routing.Routing
+import io.ktor.routing.Route
 import org.worldcubeassociation.tnoodle.server.RouteHandler
 
 object StaticContentHandler : RouteHandler {
-    override fun install(router: Routing) {
+    override fun install(router: Route) {
         router.static {
             resource("favicon.ico", "wca/favicon.ico")
             resource("robots.txt", "wca/tnoodle-ui/robots.txt")
